+++ replicante/src/components/discovery/metrics.rs
@@ -22,9 +22,9 @@ lazy_static! {
         .buckets(vec![0.25, 0.5, 1.0, 2.5, 5.0, 10.0, 20.0, 40.0])
     )
     .expect("Failed to create DISCOVERY_DURATION");
-    pub static ref DISCOVERY_ERRORS: Counter = Counter::with_opts(Opts::new(
-        "replicore_discovery_errors",
-        "Number of errors during agent discovery"
+    pub static ref DISCOVERY_LOOP_ERRORS: Counter = Counter::with_opts(Opts::new(
+        "replicore_discovery_loop_errors",
+        "Number of errors during agent discovery loops"
     ))
     .expect("Failed to create DISCOVERY_ERRORS");
     pub static ref DISCOVERY_SNAPSHOT_TRACKED_CLUSTERS: Gauge = Gauge::with_opts(Opts::new(
@@ -44,8 +44,8 @@ pub fn register_metrics(logger: &Logger, registry: &Registry) {
     if let Err(error) = registry.register(Box::new(DISCOVERY_DURATION.clone())) {
         debug!(logger, "Failed to register DISCOVERY_DURATION"; "error" => ?error);
     }
-    if let Err(error) = registry.register(Box::new(DISCOVERY_ERRORS.clone())) {
-        debug!(logger, "Failed to register DISCOVERY_ERRORS"; "error" => ?error);
+    if let Err(error) = registry.register(Box::new(DISCOVERY_LOOP_ERRORS.clone())) {
+        debug!(logger, "Failed to register DISCOVERY_LOOP_ERRORS"; "error" => ?error);
     }
     if let Err(error) = registry.register(Box::new(DISCOVERY_SNAPSHOT_TRACKED_CLUSTERS.clone())) {
         debug!(logger, "Failed to register DISCOVERY_SNAPSHOT_TRACKED_CLUSTERS"; "error" => ?error);