+++ replicante/src/metrics.rs
@@ -1,4 +1,5 @@
 use lazy_static::lazy_static;
+use prometheus::Gauge;
 use prometheus::GaugeVec;
 use prometheus::Opts;
 use prometheus::Registry;
@@ -14,6 +15,11 @@ lazy_static! {
         &["component", "type"]
     )
     .expect("Failed to create COMPONENTS_ENABLED gauge");
+    pub static ref UPDATE_AVAILABLE: Gauge = Gauge::new(
+        "replicore_updateable",
+        "Set to 1 when an updateded version is available (checked at start only)",
+    )
+    .expect("Failed to create UPDATE_AVAILABLE gauge");
     pub static ref WORKERS_ENABLED: GaugeVec = GaugeVec::new(
         Opts::new(
             "replicore_workers_enabled",
@@ -30,6 +36,9 @@ lazy_static! {
 pub fn register_metrics(logger: &Logger, registry: &Registry) {
     if let Err(error) = registry.register(Box::new(COMPONENTS_ENABLED.clone())) {
         debug!(logger, "Failed to register COMPONENTS_ENABLED"; "error" => ?error);
+    }
+    if let Err(error) = registry.register(Box::new(UPDATE_AVAILABLE.clone())) {
+        debug!(logger, "Failed to register UPDATE_AVAILABLE"; "error" => ?error);
     }
     if let Err(error) = registry.register(Box::new(WORKERS_ENABLED.clone())) {
         debug!(logger, "Failed to register WORKERS_ENABLED"; "error" => ?error);